backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pandas==2.2.3
numpy==2.2.1
python-multipart==0.0.20
httpx==0.28.1
python-dotenv==1.0.1
google-genai>=1.0.0
